bad vae

+++ gaussian_process.py
@@ -10,10 +10,10 @@ def kernal_SE(tau,signal_var,noise_var):
         data_length = data_times.shape[0]
         data_times_square = torch.outer(torch.square(data_times),torch.ones(data_length))
         K = signal_var * torch.exp(-1*(data_times_square + data_times_square.T - 2 * torch.outer(data_times,data_times))/(2*tau**2))+noise_var*torch.diag(torch.ones(data_length))
-        return K
+        return K.float()
     return kernal
 
 def zero_mean():
     def zero(data_times):
-        return torch.zeros(data_times.shape[0],dtype=float)
+        return torch.zeros(data_times.shape[0]).float()
     return zero
